Capacitance panel, slider, and caption
